Create dataModel file for dataModel functions
