+++ src/scripts.js
@@ -15,6 +15,13 @@ import {
   displayAverageSteps,
 } from './domUpdates';
 
+import {
+  getUserData,
+  getAvgSteps,
+  getAvgDailyOunces,
+  getOzByDay,
+} from './dataModel';
+
 // importing functionality from .domUpdates
 // ./ is a relative path, the module is in the same directory as the current module file
 
@@ -25,40 +32,7 @@ var welcomeHeading = document.querySelector('.welcome-heading');
 var activityContainer = document.querySelector('.average-steps')
 var friendList = document.querySelector('.friend-list')
 
-const getUserData = ((userId) => {
-    let filteredById = userData.users.find(user => user.id === userId);
-    return filteredById
-});
 
-const getAvgSteps = () => {
-  let sumOfSteps = userData.users.reduce((sum, user) => {
-    sum += user.dailyStepGoal
-    return sum
-  }, 0)
-  return sumOfSteps / userData.users.length
-}
-
-const getAvgDailyOunces = (id) => {
-  const usersDailyHydrationLog = userHydrationData.hydrationData
-  let numOfEntries = []
-  const userHydrationStats = usersDailyHydrationLog.reduce((accum, userObj) => {
-    if (userObj.userID === id) {
-      numOfEntries.push(userObj.userID)
-      accum += userObj.numOunces
-    }
-    return accum
-  }, 0)
-  return userHydrationStats / numOfEntries.length
-}
-
-///////////return specific day fluid ounces//////////////////
-const getOzByDay = (id, day) => {
-  const usersDailyHydrationLog = userHydrationData.hydrationData
-  const usersDailyOz = usersDailyHydrationLog.find(log => log.userID === id && log.date === day)
-  if(usersDailyOz) {
-    return console.log('look here', usersDailyOz.numOunces)
-  }
-}
 
 // event listener:
 window.addEventListener('load', () => {
@@ -74,9 +48,4 @@ export {
   activityContainer,
   welcomeHeading,
   friendList,
-  // functions
-  getUserData,
-  getAvgSteps,
-  getAvgDailyOunces,
-  getOzByDay
 }
